requirements.txt
cloudpickle==3.1.1
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.2
scikit_learn==1.6.1
streamlit==1.42.2
statsmodels==0.14.4
pillow==11.0.0
huggingface-hub==0.35.3